+++ core/model/group_stats.go
@@ -1,3 +1,17 @@
+// Copyright 2022 Board of Trustees of the University of Illinois.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package model
 
 // GroupStats wraps group statistics aggregation result
@@ -9,3 +23,14 @@ type GroupStats struct {
 	RejectedCount   int `json:"rejected_count" bson:"rejected_count"`
 	AttendanceCount int `json:"attendance_count" bson:"attendance_count"`
 } //@name GroupStats
+
+// StatsFilter is a filter for event statistics.
+type StatsFilter struct {
+	BaseFilter GroupsFilter            `json:"base_filter" bson:"base_filter"`
+	SubFilters map[string]GroupsFilter `json:"sub_filters" bson:"sub_filters"`
+} //@name StatsFilter
+
+// StatsResult represents the result of an event statistics query.
+type StatsResult struct {
+	Stats map[string]int64 `json:"stats" bson:"stats"`
+} //@name StatsResult